+++ DM2122_Framework/Application/Source/Characters.h
@@ -10,21 +10,21 @@ public:
 	Characters();
 	Characters(Scene*, const string& name);
 	Characters(Scene*, const string& name, Vector3 &postion);
-	Characters(Scene*, const string& name, Vector3 &postion, const unsigned &maxHealth);
-	Characters(Scene*, const string& name, Vector3 &postion, const unsigned &maxHealth, const unsigned &maxStamina);
+	Characters(Scene*, const string& name, Vector3 &postion, const double &maxHealth);
+	Characters(Scene*, const string& name, Vector3 &postion, const double &maxHealth, const double &maxStamina);
 	//Characters(Scene*, const string& name, Vector3 &postion, const unsigned &maxHealth, const unsigned &maxStamina,Inventory inventory);
 	virtual ~Characters();
 
-	virtual const float getCurrentHealth();
-	virtual const float getMaxHealth();
-	virtual const float getCurrentStamina();
-	virtual const float getMaxStamina();
+	virtual const int getCurrentHealth();
+	virtual const int getMaxHealth();
+	virtual const int getCurrentStamina();
+	virtual const int getMaxStamina();
 
-	virtual void setMaxHealth(const float newMaxHealth);
-	virtual void setMaxStamina(const float newMaxStamina);
+	virtual void setMaxHealth(const double newMaxHealth);
+	virtual void setMaxStamina(const double newMaxStamina);
 
-	virtual void updateHealth(float);
-	virtual void updateStamina(float);
+	virtual void updateHealth(double);
+	virtual void updateStamina(double);
 
 	virtual void setDefault();
 	virtual bool anyInteraction();
@@ -33,7 +33,7 @@ public:
 	Inventory inventory_;
 private:
 
-	float health_, maxHealth_, stamina_, maxStamina_;
+	double health_, maxHealth_, stamina_, maxStamina_;
 	
 };
